replatformingStruts

+++ src/main/java/org/superbiz/struts/UserService.java
@@ -21,9 +21,11 @@ import java.util.List;
 
 public interface UserService {
 
+
+
     public void add(User user);
 
-    public User find(int id);
+    public User find(long id);
 
     public List<User> findAll();
 }
